Renamed the config package to internal in the lutron binding. See openhab/static-code-analysis#100
